Add files via upload
The first in this series to operationalize Red Canary's 2022 Threat Report. The report is incredibly detailed and well executed. Our goal here is to inform defenders on additional strategies for using it. This entry covers the top techniques/sub-techniques used. 
The Report: https://redcanary.com/threat-detection-report/

Red canary provides atomic  test examples you can easily run. This plan goes deeper into those techniques with 30 tests.  
For sub-techniques with multiple tests, the external resources are not necessarily test specific, though some are, they are to help you have a deeper understanding of the technique. https://attack.mitre.org/ has a collection of references at the bottom of each technique/sub -technique.
Most of the tools used are part of regular operations. While adversaries use custom/commercial/open source tooling, many of the ways they can pivot and dig deeper into your network are with tools native to Windows/sysinternals.  We chose less complex tests to ensure usability, especially for first time ART users. While the tests chosen are simpler to execute, they are still valuable for security teams of all levels. If a test is recommended in the report, and we also added it to our plan, it is a test we highly recommend. 

If this is your first time emulating or using Atomics, the good people at Scythe put together a very user friendly guide for operationalizing the report. It requires a copy and paste, and will provide a real feel for how atomics work. 
https://www.scythe.io/library/operationalizing-red-canarys-2022-threat-detection-report
Jorge Orchilles 🦄@jorgeorchilles & Christopher Peacock 🦄 @SecurePeacock

Be sure to check out the main read.me for resources on Atomic Red Team and Invoke-Atomic.
It is our goal to maintain accuracy, credit authors, and serve the public.
If you find any mistakes or have suggestions for this library entry, please contact us on twitter @AptEmulations. 
Take care. 
-DFIRSEC Team.